+++ src/promptflow-devkit/promptflow/_sdk/entities/_flow/dag.py
@@ -115,7 +115,7 @@ class Flow(FlowBase):
 
     # region overrides:
     def _init_executable(self, tuning_node=None, variant=None):
-        from promptflow._sdk._submitter import variant_overwrite_context
+        from promptflow._sdk._orchestrator import variant_overwrite_context
         from promptflow.contracts.flow import Flow as ExecutableFlow
 
         if not tuning_node and not variant:
@@ -143,7 +143,7 @@ class Flow(FlowBase):
 
     def invoke(self, inputs: dict) -> "LineResult":
         """Invoke a flow and get a LineResult object."""
-        from promptflow._sdk._submitter import TestSubmitter
+        from promptflow._sdk._orchestrator import TestSubmitter
 
         if self.language == FlowLanguage.CSharp:
             # TODO 3033484: we shouldn't use context manager here for stream_output, as resource need to be released
